[posix] print connect error
